# docs/openapi.yaml
openapi: 3.1.0
info:
  title: codex-browser buyer API
  version: 0.1.0
  summary: HTTP API сервиса buyer.
  description: |
    Документ фиксирует текущий MVP-контракт HTTP API сервиса `buyer`.

    `buyer` принимает задачи покупки, ведет сессии и отправляет callback-события
    в `callback_url`. Контракт исходящих callback-вызовов вынесен в
    `docs/callbacks.openapi.yaml`.
servers:
  - url: http://localhost:8000
    description: Локальный buyer API.
tags:
  - name: Buyer API
    description: Основной API сервиса buyer.

paths:
  /healthz:
    get:
      tags:
        - Buyer API
      operationId: healthz
      summary: Проверить доступность buyer.
      responses:
        '200':
          description: Buyer отвечает.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/tasks:
    post:
      tags:
        - Buyer API
      operationId: createBuyerTask
      summary: Создать сессию покупки.
      description: |
        Запускает асинхронный сценарий покупки. Ответ возвращается сразу после
        создания сессии и постановки runner в работу.

        Дальнейший прогресс приходит callback-событиями на `callback_url`.
        Если `callback_url` не передан, buyer использует `MIDDLE_CALLBACK_URL`.
        `start_url` должен быть публичным http/https URL без userinfo и без
        loopback/private/link-local/metadata hostname или IP. Task-provided
        `callback_url` должен быть публичным https URL или совпадать с
        доверенным callback URL из `TRUSTED_CALLBACK_URLS`; default
        `MIDDLE_CALLBACK_URL` тоже считается доверенным. `callback_url`,
        `MIDDLE_CALLBACK_URL` и entries `TRUSTED_CALLBACK_URLS` не должны
        содержать query string или fragment.
        Если callback receiver требует секрет, передавайте его отдельным
        `callback_token`: buyer отправит его как `X-Eval-Callback-Token`
        и не сохранит в persistent state.
        Формат callback-событий описан в `docs/callbacks.openapi.yaml`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
            examples:
              litres:
                summary: Покупка книги на Litres.
                value:
                  task: Ищи книгу 1984, доведи до шага оплаты через SberPay без финального платежа.
                  start_url: https://www.litres.ru/
                  callback_url: https://middle.example.com/callbacks
                  metadata:
                    source: openclaw
      responses:
        '201':
          description: Сессия создана и переведена в работу.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
        '409':
          description: Нельзя создать новую активную сессию из-за runtime-лимита.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /v1/sessions:
    get:
      tags:
        - Buyer API
      operationId: listBuyerSessions
      summary: Получить список сессий.
      responses:
        '200':
          description: Список сессий.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionView'

  /v1/sessions/{session_id}:
    get:
      tags:
        - Buyer API
      operationId: getBuyerSession
      summary: Получить детальное состояние сессии.
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '200':
          description: Состояние сессии и сохраненные callback-события.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          description: Сессия не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /v1/sessions/{session_id}/stop:
    post:
      tags:
        - Buyer API
      operationId: stopBuyerSession
      summary: Остановить активную сессию покупки.
      description: |
        Останавливает активную сессию (`created`, `running`, `waiting_user`) как
        штатную failed-финализацию. Buyer отправляет финальный callback
        `scenario_finished` с `status=failed` и
        `reason_code=session_stopped_by_operator`, отменяет runtime task и
        завершает активный `codex exec` subprocess, если он уже запущен.

        Для уже терминальных сессий (`completed`, `failed`) endpoint
        идемпотентен: состояние не меняется, а `accepted=false`.
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionStopRequest'
            examples:
              operatorStop:
                value:
                  reason: Оператор остановил сценарий.
      responses:
        '200':
          description: Запрос остановки обработан.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStopResponse'
        '404':
          description: Сессия не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  /v1/replies:
    post:
      tags:
        - Buyer API
      operationId: submitBuyerReply
      summary: Передать ответ пользователя в ожидающую сессию.
      description: |
        Используется после callback-события `ask_user`. `reply_id` должен
        совпадать с `payload.reply_id` из актуального события ожидания.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionReplyRequest'
            examples:
              reply:
                value:
                  session_id: 8b69d854-2d5d-4fb3-9946-d09736c8a461
                  reply_id: 3c5c7077-d7af-4f55-9802-f9e3d71c8707
                  message: Да, продолжай с этим вариантом.
      responses:
        '200':
          description: Ответ принят и сессия продолжила выполнение.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionReplyResponse'
        '404':
          description: Сессия не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Сессия не ожидает этот ответ или не имеет активного runner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

components:
  parameters:
    SessionIdPath:
      name: session_id
      in: path
      required: true
      schema:
        type: string
      description: Идентификатор сессии buyer.

  responses:
    ValidationError:
      description: Ошибка валидации FastAPI/Pydantic.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'

  schemas:
    JsonValue:
      description: Любое JSON-значение.

    JsonObject:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/JsonValue'

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          const: ok

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          description: Текст или структурированное описание ошибки.
          oneOf:
            - type: string
            - type: object
              additionalProperties:
                $ref: '#/components/schemas/JsonValue'
            - type: array
              items:
                $ref: '#/components/schemas/JsonValue'

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'

    ValidationErrorItem:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
        input:
          $ref: '#/components/schemas/JsonValue'

    SessionStatus:
      type: string
      enum:
        - created
        - running
        - waiting_user
        - completed
        - failed

    PlaywrightStorageState:
      type: object
      description: Playwright `storageState` для восстановления браузерного auth-контекста.
      properties:
        cookies:
          type: array
          items:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/JsonValue'
        origins:
          type: array
          items:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/JsonValue'
      additionalProperties:
        $ref: '#/components/schemas/JsonValue'

    TaskAuthPayload:
      type: object
      description: |
        Опциональный inline auth-пакет. В v1 поддерживается SberId.

        Если пакет передан, он имеет приоритет над внешним cookies API,
        настроенным через env. Пользовательские replies не используются
        для передачи cookies, `storageState`, localStorage или tokens.
      properties:
        provider:
          type: string
          default: sberid
          description: Провайдер авторизации.
        storageState:
          oneOf:
            - $ref: '#/components/schemas/PlaywrightStorageState'
            - type: 'null'
          description: Playwright storageState для восстановления сессии.
      additionalProperties: false

    TaskCreateRequest:
      type: object
      required:
        - task
        - start_url
      properties:
        task:
          type: string
          minLength: 1
          description: Текст задачи для агента buyer.
        start_url:
          type: string
          minLength: 1
          description: |
            URL магазина для начала сценария. Разрешены только публичные
            http/https URL с hostname, без userinfo и без
            loopback/private/link-local/metadata hosts.
        callback_url:
          oneOf:
            - type: string
            - type: 'null'
          default: null
          description: |
            Куда buyer отправляет callback-события. Схема исходящих callback-вызовов
            описана в `docs/callbacks.openapi.yaml`. Если поле передано задачей,
            URL должен быть публичным https URL или совпадать с trusted callback
            allowlist `TRUSTED_CALLBACK_URLS`. Если поле не передано, используется
            `MIDDLE_CALLBACK_URL`. Query string и fragment запрещены для
            task-provided, default и trusted callback URL.
        callback_token:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
          default: null
          description: |
            Ephemeral token для callback receiver. Buyer отправляет его в
            заголовке `X-Eval-Callback-Token` и не сохраняет в persistent state.
        metadata:
          $ref: '#/components/schemas/JsonObject'
          default: {}
        auth:
          oneOf:
            - $ref: '#/components/schemas/TaskAuthPayload'
            - type: 'null'
          default: null
      additionalProperties: false

    TaskCreateResponse:
      type: object
      required:
        - session_id
        - status
        - novnc_url
      properties:
        session_id:
          type: string
        status:
          $ref: '#/components/schemas/SessionStatus'
        novnc_url:
          type: string

    SessionReplyRequest:
      type: object
      required:
        - session_id
        - reply_id
        - message
      properties:
        session_id:
          type: string
        reply_id:
          type: string
        message:
          type: string
          minLength: 1
      additionalProperties: false

    SessionReplyResponse:
      type: object
      required:
        - session_id
        - accepted
        - status
      properties:
        session_id:
          type: string
        accepted:
          type: boolean
        status:
          $ref: '#/components/schemas/SessionStatus'

    SessionStopRequest:
      type: object
      properties:
        reason:
          oneOf:
            - type: string
              minLength: 1
            - type: 'null'
          default: null
          description: Причина остановки, которая попадет в `scenario_finished.payload.message`.
      additionalProperties: false

    SessionStopResponse:
      type: object
      required:
        - session_id
        - accepted
        - status
      properties:
        session_id:
          type: string
        accepted:
          type: boolean
          description: '`true`, если сессия была активной и остановка изменила состояние.'
        status:
          $ref: '#/components/schemas/SessionStatus'

    SessionView:
      type: object
      required:
        - session_id
        - status
        - start_url
        - callback_url
        - novnc_url
        - created_at
        - updated_at
      properties:
        session_id:
          type: string
        status:
          $ref: '#/components/schemas/SessionStatus'
        start_url:
          type: string
        callback_url:
          type: string
        novnc_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        waiting_reply_id:
          oneOf:
            - type: string
            - type: 'null'
          default: null
        last_error:
          oneOf:
            - type: string
            - type: 'null'
          default: null

    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/SessionView'
        - type: object
          required:
            - events
          properties:
            events:
              type: array
              items:
                $ref: './callbacks.openapi.yaml#/components/schemas/CallbackEventEnvelope'